+++ packages/create/templates/hello-knex/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@dbos-inc/dbos-sdk": "file:../../../..",
+    "express": "^4.21.2",
     "knex": "3.1.0"
   }
 }
